Chore: Finish writing tests for getting all events. Update README to show the command for retrieving all events using Postman. Feat: Write code to pass the tests.
