+++ README.md
@@ -14,6 +14,7 @@
 * Send a DELETE request to 127.0.0.1:8000/centers/1
 * Send a GET request to 127.0.0.1:8000/centers
 * Send a GET request to 127.0.0.1:8000/centers/1
+* Send a GET request to 127.0.0.1:8000/events
 
 ## Tests
 > To run tests, use `npm test`
