+++ server/controllers/events.js
@@ -1,26 +1,5 @@
-const events = [
-  {
-    id: '1',
-    name: 'kachi\'s event',
-    date: '2011-11-11',
-    time: '08:00',
-    centerId: '1'
-  },
-  {
-    id: '2',
-    name: 'kachi\'s second event',
-    date: '2011-11-11',
-    time: '08:00',
-    centerId: '3'
-  },
-  {
-    id: '3',
-    name: 'kachi\'s event',
-    date: '2011-11-11',
-    time: '08:00',
-    centerId: '2'
-  }
-];
+import events from '../models/events';
+
 const newEvents = [];
 
 module.exports = {
@@ -65,5 +44,8 @@ module.exports = {
         message: 'Resource deleted successfully'
       });
     }
+  },
+  get(req, res) {
+    res.status(200).send(events);
   }
 };
