+++ server/models/events.js
@@ -1,0 +1,25 @@
+const events = [
+  {
+    id: '1',
+    name: 'kachi\'s event',
+    date: '2011-11-11',
+    time: '08:00',
+    centerId: '1'
+  },
+  {
+    id: '2',
+    name: 'kachi\'s second event',
+    date: '2011-11-11',
+    time: '08:00',
+    centerId: '3'
+  },
+  {
+    id: '3',
+    name: 'kachi\'s event',
+    date: '2011-11-11',
+    time: '08:00',
+    centerId: '2'
+  }
+];
+
+module.exports = events;
